chore 🔥: move examples and split models
