+++ examples/simulated_cnn_cifar10/fedavg.py
