+++ examples/simulated_cnn_cifar10/single.py
