+++ examples/simulated_cnn_mnist/fedavg.py
