+++ fedhf/model/criterion/__init__.py
@@ -1,0 +1,28 @@
+#!/usr/bin/env python
+# -*- encoding: utf-8 -*-
+""" 
+@File    :   fedhf\model\criterion\__init__.py 
+@Time    :   2022-01-24 11:49:40 
+@Author  :   Bingjie Yan 
+@Email   :   bj.yan.pa@qq.com 
+@License :   Apache License 2.0 
+"""
+
+import torch
+import torch.nn as nn
+import torch.optim as optim
+
+criterion_factory = {
+    'l1': nn.L1Loss,
+    'mse': nn.MSELoss,
+    'ce': nn.CrossEntropyLoss,
+    'bce': nn.BCELoss,
+}
+
+
+def build_criterion(criter_name: str):
+    if criter_name not in criterion_factory.keys():
+        raise ValueError(f'Unknown criterion name: {criter_name}')
+
+    cirter = criterion_factory[criter_name]
+    return cirter
